src/product_embeddings/generate_embedding/requirements.txt
functions-framework==3.*
google-genai==1.*
google-cloud-logging==3.11.*
google-cloud-bigquery==3.29.*
